# server/openapi.yaml
openapi: '3.1.0'
info:
  title: Safedify AI — HSE Management Platform API
  version: '1.0.0'
  description: |
    REST API for the Safedify AI Health, Safety & Environment management platform.
    All authenticated endpoints require `Authorization: Bearer <token>` header.
  contact:
    name: Safedify AI
    url: https://github.com/Thundhai/Safedify-AI
servers:
  - url: /api
    description: Default (same-origin in production, proxied in dev)
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    User:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        email: { type: string, format: email }
        role: { type: string }
        tier: { type: string }
        avatar: { type: string }
...
    SearchResult:
      type: object
      properties:
        query: { type: string }
        total: { type: integer }
        results:
          type: array
          items:
            type: object
            properties:
              id: { type: string }
              type: { type: string }
              title: { type: string }
              snippet: { type: string }
              rank: { type: number }
  parameters:
    PageParam:
      name: page
      in: query
      schema: { type: integer, default: 1 }
    LimitParam:
      name: limit
      in: query
      schema: { type: integer, default: 50, maximum: 200 }
security:
  - BearerAuth: []
paths:
...
  /auth/login:
    post:
      tags: [Auth]
      summary: Login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string }
      responses:
        '200':
          description: JWT token and user info
          content:
            application/json:
              schema:
                type: object
                properties:
                  token: { type: string }
                  user: { $ref: '#/components/schemas/User' }
        '401':
          description: Invalid credentials
  /auth/register:
    post:
      tags: [Auth]
      summary: Register new user
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, email, password]
              properties:
                name: { type: string }
                email: { type: string, format: email }
                password: { type: string, minLength: 8 }
                role: { type: string, enum: [Worker, HSE Supervisor] }
      responses:
        '201':
          description: Account created
        '409':
          description: Email already registered
...
  /search:
    get:
      tags: [Search]
      summary: Full-text search across all entities
      parameters:
        - name: q
          in: query
          required: true
          schema: { type: string, minLength: 2 }
          description: Search query
        - name: type
          in: query
          schema: { type: string }
          description: Comma-separated entity types to search (incidents,observations,actions,permits,documents)
        - $ref: '#/components/parameters/LimitParam'
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SearchResult' }
...
  /uploads:
    post:
      tags: [Uploads]
      summary: Upload file (base64)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                data: { type: string, description: 'Base64 data URI' }
                files: { type: array, items: { type: string }, description: 'Multiple base64 URIs' }
      responses:
        '201':
          description: Upload result
...
  /audit-logs:
    get:
      tags: [Admin]
      summary: List audit logs (Admin only)
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - name: action
          in: query
          schema: { type: string }
        - name: entity_type
          in: query
          schema: { type: string }
        - name: from
          in: query
          schema: { type: string, format: date-time }
        - name: to
          in: query
          schema: { type: string, format: date-time }
      responses:
        '200': { description: Paginated audit logs }
...